LogWriter: Refactor to own package
Refactor LogChunkWriter into its own package.
